fix(overviewmap): was not loading on first instance due to async from baselayer
