+++ src/features/navigation/controls/index.ts
@@ -3,11 +3,11 @@ import { Collection } from "ol";
 import { useTranslation } from "@/i18n";
 import { translateLayerSwitcherControl } from "@/constants/communities/utils";
 import DrawingControl from "./DrawingControl";
-import useGetToolsControls from "./ToolsControl";
+import useToolsControls from "./ToolsControl";
 
 const useGetMapControls = (): Collection<Control> | Control[] | undefined => {
     const drawingControl = DrawingControl();
-    const toolsControls = useGetToolsControls();
+    const toolsControls = useToolsControls();
     const { t } = useTranslation({ useGetMapControls });
     setTimeout(() => {
         translateLayerSwitcherControl(t);
